Rename apprise.py to stop overriding library & Fix URL addition error
