+++ wolnut/state.py
@@ -3,7 +3,7 @@ import os
 from typing import Dict, Any
 import logging
 import time
-from wolnut.apprise import apprise_notifier
+from wolnut.apprise_logging import apprise_notifier
 
 logger = apprise_notifier("wolnut")
 logger.setLevel(logging.INFO)
